src/requirements.txt
grpcio
grpcio-tools
protobuf
pymongo
Flask
flask-cors
Flask-Session
pandas
scikit-learn